Agent_Analista_de_excel/requirements.txt
crewai-tools[mcp]
crewai
fastmcp
mcpadapt
python-dotenv
setuptools
streamlit
langchain-openai
mcp[cli]
nest-asyncio